requirements.txt
django==5.2.6
channels==4.3.1
daphne==4.2.1
channels_redis==4.3.0
psycopg2==2.9.10
djangorestframework==3.16.0